moved userdata parsing utility classes
