+++ src/utility/user_data/parsing/getClassesForWeekday.js
