+++ src/utility/user_data/parsing/getTimetableDataForWeek.js
